Fix type mismatch for GetVirtualSizeWith<Descendants|Ancerstors>
Summary:
These methods are declared to return an `uint64_t` but there definition
returns a `size_t`, which breaks the build on 32 bit platforms (and OSX
for some reason).

Test Plan:
  ninja check
Run the Gitian builds.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5024

+++ src/txmempool.cpp
@@ -50,15 +50,15 @@ size_t CTxMemPoolEntry::GetTxVirtualSize() const {
     return GetVirtualTransactionSize(nTxSize, sigOpCount);
 }
 
-size_t CTxMemPoolEntry::GetVirtualSizeWithDescendants() const {
+uint64_t CTxMemPoolEntry::GetVirtualSizeWithDescendants() const {
     // note this is distinct from the sum of descendants' individual virtual
     // sizes, and may be smaller.
     return GetVirtualTransactionSize(nSizeWithDescendants,
                                      nSigOpCountWithDescendants);
 }
 
-size_t CTxMemPoolEntry::GetVirtualSizeWithAncestors() const {
-    // note this is distinct from the sum of descendants' individual virtual
+uint64_t CTxMemPoolEntry::GetVirtualSizeWithAncestors() const {
+    // note this is distinct from the sum of ancestors' individual virtual
     // sizes, and may be smaller.
     return GetVirtualTransactionSize(nSizeWithAncestors,
                                      nSigOpCountWithAncestors);
